% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RDI.R
\name{calculate_RDI}
\alias{calculate_RDI}
\title{Calculate RNA Degradation Index - RDI}
\usage{
calculate_RDI(fq_file1, fq_file2, index_path, num_threads, output_dir)
}
\arguments{
\item{fq_file1}{Path to the first RNAseq fq.gz file}

\item{fq_file2}{Path to the second RNAseq fq.gz file}

\item{index_path}{Path to the Kallisto index}

\item{num_threads}{Number of threads to use}

\item{output_dir}{Path to the output directory}
}
\value{
Outputs two files: RDI for each transcript, and the RDI for the sample
}
\description{
Function to calculate RDI
}
\examples{
calculate_RDI(fq_file1, fq_file2, index_path, num_threads, output_dir)
}
