% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_index.R
\name{create_index}
\alias{create_index}
\title{Create kallisto Index}
\usage{
create_index(fasta_file, output_prefix)
}
\arguments{
\item{fasta_file}{Path to the input the original reference transcript, in fasta format}

\item{output_prefix}{Path to the generated reference transcript}
}
\value{
Outputs the kallisto index that will be used for calculating RDI
}
\description{
Function to create the reference transcript and the indexed reference transcript
}
\examples{
create_index(fasta_file, output_prefix)
}
